refactor: polish tab rule handling and UI
