+++ extension/src/locales/en.json
@@ -38,7 +38,7 @@
         "title": "Excluded Resource Types"
       },
       "excludedTabIds": {
-        "description": "The rule will not match requests from the selected tabs. Tab bindings are registered as session rules.",
+        "description": "The rule will not match requests from the selected tabs. Tab bindings become invalid after the browser restarts.",
         "title": "Excluded Tabs"
       },
       "initiatorDomains": {
@@ -78,7 +78,7 @@
         "title": "Simple Redirect URL"
       },
       "tabIds": {
-        "description": "The rule will only match requests from the selected tabs. Tab bindings are registered as session rules.",
+        "description": "The rule will only match requests from the selected tabs. Tab bindings become invalid after the browser restarts.",
         "title": "Tabs"
       },
       "urlFilter": {
@@ -223,11 +223,11 @@
     },
     "selectPlaceholder": "Select {name}...",
     "tabIds": {
-      "description": "The rule only matches requests originating from a selected tab. Tab IDs expire when the browser session ends.",
+      "description": "The rule only matches requests originating from a selected tab. Selected tabs become invalid after the browser restarts.",
       "quickSelect": "Quickly replace selected tabs",
       "selectPlaceholder": "Select tabs...",
       "title": "Tabs",
-      "unavailableTab": "Tab is unavailable in this browser session",
+      "unavailableTab": "Tab is unavailable because it was closed or the browser was restarted",
       "untitledTab": "Untitled tab",
       "useCurrentGroup": "Use current group",
       "useCurrentTab": "Use current tab",
